trivial: Move algo/* to algo/x16r
