+++ src/Makefile.am
@@ -83,49 +83,49 @@ endif
 AVIAN_CORE_H = \
   addrdb.h \
   addressindex.h \
-  algo/sph_groestl.h \
-  algo/sph_types.h \
-  algo/sph_blake.h \
-  algo/sph_bmw.h \
-  algo/sph_jh.h \
-  algo/sph_keccak.h \
-  algo/sph_skein.h \
-  algo/sph_luffa.h \
-  algo/sph_cubehash.h \
-  algo/sph_simd.h \
-  algo/sph_echo.h \
-  algo/sph_shavite.h \
-  algo/sph_sha2.h \
-  algo/sph_haval.h \
-  algo/sph_tiger.h \
-  algo/lyra2.h \
-  algo/sponge.h \
-  algo/gost_streebog.h \
-  algo/groestl.c \
-  algo/blake.c \
-  algo/bmw.c \
-  algo/cubehash.c \
-  algo/echo.c \
-  algo/jh.c \
-  algo/keccak.c \
-  algo/luffa.c \
-  algo/shavite.c \
-  algo/simd.c \
-  algo/skein.c \
-  algo/hamsi.c \
-  algo/whirlpool.c \
-  algo/shabal.c \
-  algo/fugue.c \
-  algo/sha2.c \
   algo/extra.c \
   algo/extra.h \
-  algo/sph_sha2big.c \
-  algo/haval.c \
-  algo/tiger.cpp \
-  algo/lyra2.cpp \
-  algo/sponge.cpp \
-  algo/sph_sha2.c \
-  algo/gost_streebog.c \
+  algo/x16r/sph_groestl.h \
+  algo/x16r/sph_types.h \
+  algo/x16r/sph_blake.h \
+  algo/x16r/sph_bmw.h \
+  algo/x16r/sph_jh.h \
+  algo/x16r/sph_keccak.h \
+  algo/x16r/sph_skein.h \
+  algo/x16r/sph_luffa.h \
+  algo/x16r/sph_cubehash.h \
+  algo/x16r/sph_simd.h \
+  algo/x16r/sph_echo.h \
+  algo/x16r/sph_shavite.h \
+  algo/x16r/sph_sha2.h \
+  algo/x16r/sph_haval.h \
+  algo/x16r/sph_tiger.h \
+  algo/x16r/lyra2.h \
+  algo/x16r/sponge.h \
+  algo/x16r/gost_streebog.h \
+  algo/x16r/groestl.c \
+  algo/x16r/blake.c \
+  algo/x16r/bmw.c \
+  algo/x16r/cubehash.c \
+  algo/x16r/echo.c \
+  algo/x16r/jh.c \
+  algo/x16r/keccak.c \
+  algo/x16r/luffa.c \
+  algo/x16r/shavite.c \
+  algo/x16r/simd.c \
+  algo/x16r/skein.c \
+  algo/x16r/hamsi.c \
+  algo/x16r/whirlpool.c \
+  algo/x16r/shabal.c \
+  algo/x16r/fugue.c \
+  algo/x16r/sha2.c \
+  algo/x16r/sph_sha2big.c \
+  algo/x16r/haval.c \
+  algo/x16r/tiger.cpp \
+  algo/x16r/lyra2.cpp \
+  algo/x16r/sponge.cpp \
+  algo/x16r/sph_sha2.c \
+  algo/x16r/gost_streebog.c \
   algo/minotaurx/blake2.h \
   algo/minotaurx/blake2-impl.h \
   algo/minotaurx/blake2s-ref.c \
