+++ src/algo/extra.c
@@ -1,3 +1,5 @@
+// Copyright (c) 2017-2019 The Raven Core developers
+
 #include <stdint.h>
 #include <string.h>
