+++ src/algo/extra.h
@@ -1,3 +1,5 @@
+// Copyright (c) 2017-2019 The Raven Core developers
+
 uint32_t be32dec(uint32_t *pp);
 void be32enc(uint32_t *pp, uint32_t x);
 uint32_t le32dec(uint32_t *pp);
