+++ src/algo/x16r/blake.c
