+++ src/algo/x16r/bmw.c
