+++ src/algo/x16r/cubehash.c
