+++ src/algo/x16r/echo.c
