+++ src/algo/x16r/fugue.c
