+++ src/algo/x16r/gost_streebog.c
