+++ src/algo/x16r/hamsi.c
