+++ src/algo/x16r/hamsi_helper.c
