+++ src/algo/x16r/haval.c
