+++ src/algo/x16r/haval_helper.c
