+++ src/algo/x16r/lyra2.cpp
