+++ src/algo/x16r/md_helper.c
