+++ src/algo/x16r/shabal.c
