+++ src/algo/x16r/simd.c
