+++ src/algo/x16r/skein.c
