+++ src/algo/x16r/sph_cubehash.h
