+++ src/algo/x16r/sph_hamsi.h
