+++ src/algo/x16r/sph_simd.h
