+++ src/algo/x16r/sph_skein.h
