+++ src/algo/x16r/sph_whirlpool.h
