+++ src/algo/x16r/sponge.cpp
