+++ src/algo/x16r/whirlpool.c
