+++ src/hash.h
@@ -1,7 +1,7 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
 // Copyright (c) 2012-2018 The Bitcoin Core developers
 // Copyright (c) 2017 The Raven Core developers
 // Copyright (c) 2021 The Avian Core developers
-// Copyright (c) 2009-2010 Satoshi Nakamoto
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -17,7 +17,7 @@
 #include "version.h"
 
 extern "C" {
-#include "algo/sph_sha2.h"
+#include "algo/x16r/sph_sha2.h"
 }
 
 #include <vector>
